refactor: remove eventAsstPlus agent and add moderator support to eventAsst when enabled
provide mod support (/mod and question follow-up) when agentConfig.moderatorSupport is defined

BREAKING CHANGE: removes eventAssistantPlus agent, impacts only active conversations
